refactor(ios): separate cloud sync failure and result policy
