+++ apps/ios/Flashcards/Flashcards/Cloud/Store/Sync/FlashcardsStore+CloudSync.swift
@@ -5,13 +5,6 @@ private enum PersistedCloudStateReconciliationOutcome {
     case stopSync
 }
 
-private let blockedCloudIdentityConflictCodes: Set<String> = [
-    "GUEST_SESSION_PLATFORM_MISMATCH",
-    "SYNC_INSTALLATION_PLATFORM_MISMATCH",
-    "SYNC_REPLICA_CONFLICT",
-    "SYNC_WORKSPACE_FORK_REQUIRED"
-]
-
 @MainActor
 extension FlashcardsStore {
     func manualCloudSyncTrigger(now: Date) -> CloudSyncTrigger {
@@ -639,118 +632,6 @@ extension FlashcardsStore {
         }
     }
 
-    /**
-     Applies sync side effects through diff-aware bootstrap and review
-     reconciliation so no-op syncs do not trigger a blocking review reload.
-     */
-    func applySyncResultWithoutBlockingReset(
-        syncResult: CloudSyncResult,
-        now: Date,
-        trigger: CloudSyncTrigger
-    ) async throws {
-        let bootstrapRefreshOutcome = try await self.refreshBootstrapSnapshotWithoutReset(now: now)
-        let didResetVolatileReviewSelection = self.resetVolatileReviewSelectionAfterLocalIdRepairIfNeeded(
-            syncResult: syncResult,
-            now: now
-        )
-        let shouldRefreshReviewState = didResetVolatileReviewSelection == false
-            && (syncResult.reviewDataChanged || bootstrapRefreshOutcome.cardsChanged)
-        let didRefreshReviewState: Bool
-        if shouldRefreshReviewState {
-            let reviewRefreshMode: ReviewRefreshMode
-            if trigger.allowsVisibleChangeBanner || syncResult.appliedPullChanges {
-                reviewRefreshMode = .backgroundReconcileWithVisibleChangeBanner
-            } else {
-                reviewRefreshMode = .backgroundReconcileSilently
-            }
-            didRefreshReviewState = try await self.refreshReviewState(
-                now: now,
-                mode: reviewRefreshMode
-            )
-            self.reconcileStrictReminders(trigger: .reviewHistoryImported, now: now)
-        } else {
-            didRefreshReviewState = didResetVolatileReviewSelection
-            if didResetVolatileReviewSelection {
-                self.reconcileStrictReminders(trigger: .reviewHistoryImported, now: now)
-            }
-        }
-        if trigger.allowsVisibleChangeBanner {
-            self.enqueueBackgroundSyncVisibleChangeBannerIfNeeded(
-                bootstrapRefreshOutcome: bootstrapRefreshOutcome
-            )
-        }
-        if bootstrapRefreshOutcome.didChange
-            || didRefreshReviewState
-            || syncResult.changedEntityTypes.contains(.mediaAsset) {
-            self.localReadVersion += 1
-        }
-        if bootstrapRefreshOutcome.homeSnapshotChanged {
-            self.requestGuestSignInAfterReviewPromptReconciliation()
-        }
-        await self.handleProgressSyncCompletion(
-            now: now,
-            syncResult: syncResult
-        )
-        self.lastSuccessfulCloudSyncAt = nowIsoTimestamp()
-        self.syncStatus = .idle
-        self.globalErrorMessage = ""
-    }
-
-    private func failureErrorAfterApplyingLocalIdRepairSideEffectsIfNeeded(
-        error: Error,
-        now: Date
-    ) async throws -> Error {
-        guard let localIdRepairFailure = error as? CloudSyncLocalIdRepairFailure else {
-            return error
-        }
-
-        try await self.applyLocalIdRepairSideEffectsAfterSyncFailure(
-            syncResult: localIdRepairFailure.syncResult,
-            now: now
-        )
-        return localIdRepairFailure.underlyingError
-    }
-
-    private func applyLocalIdRepairSideEffectsAfterSyncFailure(
-        syncResult: CloudSyncResult,
-        now: Date
-    ) async throws {
-        let bootstrapRefreshOutcome = try await self.refreshBootstrapSnapshotWithoutReset(now: now)
-        let didResetVolatileReviewSelection = self.resetVolatileReviewSelectionAfterLocalIdRepairIfNeeded(
-            syncResult: syncResult,
-            now: now
-        )
-        if didResetVolatileReviewSelection {
-            self.reconcileStrictReminders(trigger: .reviewHistoryImported, now: now)
-        }
-        if bootstrapRefreshOutcome.didChange || didResetVolatileReviewSelection {
-            self.localReadVersion += 1
-        }
-        if bootstrapRefreshOutcome.homeSnapshotChanged {
-            self.requestGuestSignInAfterReviewPromptReconciliation()
-        }
-    }
-
-    /**
-     Local re-id recovery can invalidate volatile review filters and selections
-     that store entity ids. Reset broadly to All Cards instead of preserving
-     individual filters with fragile per-entity repair logic.
-     */
-    private func resetVolatileReviewSelectionAfterLocalIdRepairIfNeeded(
-        syncResult: CloudSyncResult,
-        now: Date
-    ) -> Bool {
-        guard syncResult.localIdRepairEntityTypes.isEmpty == false else {
-            return false
-        }
-
-        self.selectedReviewFilter = .allCards
-        self.persistSelectedReviewFilter(reviewFilter: .allCards)
-        self.startReviewLoad(reviewFilter: .allCards, now: now)
-        self.reconcileReviewNotifications(trigger: .filterChanged, now: now)
-        return true
-    }
-
     func isCloudAuthorizationError(_ error: Error) -> Bool {
         self.cloudRuntime.isCloudAuthorizationError(error)
     }
@@ -767,155 +648,6 @@ extension FlashcardsStore {
 
     func isCloudAccountDeletedError(_ error: Error) -> Bool {
         self.cloudRuntime.isCloudAccountDeletedError(error)
-    }
-
-    private func syncStatusForCloudFailure(
-        error: Error,
-        fallbackCloudState: CloudAccountState?,
-        trigger: CloudSyncTrigger
-    ) -> SyncStatus {
-        if let recoveryState = self.cloudCredentialRecoveryState {
-            return .blocked(message: localizedCloudCredentialRecoveryBlockedMessage(reason: recoveryState.reason))
-        }
-        if trigger.source == .postAuth {
-            return .idle
-        }
-
-        if let blockedMessage = self.blockedCloudIdentityConflictMessage(error: error) {
-            return .blocked(message: blockedMessage)
-        }
-
-        if fallbackCloudState == .linked || fallbackCloudState == .guest {
-            return .failed(message: Flashcards.errorMessage(error: error))
-        }
-
-        return .idle
-    }
-
-    func transitionSyncStatusForCloudFailure(error: Error) -> SyncStatus {
-        if let recoveryState = self.cloudCredentialRecoveryState {
-            return .blocked(message: localizedCloudCredentialRecoveryBlockedMessage(reason: recoveryState.reason))
-        }
-        if let blockedMessage = self.blockedCloudIdentityConflictMessage(error: error) {
-            return .blocked(message: blockedMessage)
-        }
-
-        return .failed(message: Flashcards.errorMessage(error: error))
-    }
-
-    func transitionSyncStatusForCloudFailure(error: Error, trigger: CloudSyncTrigger) -> SyncStatus {
-        if let recoveryState = self.cloudCredentialRecoveryState {
-            return .blocked(message: localizedCloudCredentialRecoveryBlockedMessage(reason: recoveryState.reason))
-        }
-        if trigger.source == .postAuth {
-            return .idle
-        }
-
-        if let blockedMessage = self.blockedCloudIdentityConflictMessage(error: error) {
-            return .blocked(message: blockedMessage)
-        }
-
-        return .failed(message: Flashcards.errorMessage(error: error))
-    }
-
-    func blockedCloudIdentityConflictMessage(error: Error) -> String? {
-        guard let syncError = error as? CloudSyncError else {
-            return nil
-        }
-        guard case .invalidResponse(let details, _) = syncError else {
-            return nil
-        }
-        guard blockedCloudIdentityConflictCodes.contains(details.code ?? "") else {
-            return nil
-        }
-        return Flashcards.errorMessage(error: error)
-    }
-
-    func captureCloudSyncFailure(
-        error: Error,
-        linkedSession: CloudLinkedSession,
-        fallbackCloudState: CloudAccountState?,
-        action: String,
-        captureContext: TechnicalErrorCaptureContext?
-    ) {
-        let diagnostics = cloudSyncFailureDiagnostics(error: error)
-        let scope = IOSObservationScope(
-            feature: .cloudSync,
-            userId: linkedSession.userId,
-            workspaceId: linkedSession.workspaceId,
-            requestId: diagnostics.requestId,
-            clientRequestId: nil,
-            sessionId: nil,
-            runId: nil,
-            cloudState: fallbackCloudState ?? self.cloudSettings?.cloudState,
-            configurationMode: linkedSession.configurationMode
-        )
-        self.markTechnicalErrorCaptured(captureContext: captureContext)
-        FlashcardsObservability.captureException(
-            .cloudSyncFailed(
-                error: error,
-                scope: scope,
-                details: CloudSyncFailureDetails(
-                    action: action,
-                    statusCode: diagnostics.statusCode,
-                    backendCode: diagnostics.backendCode,
-                    requestId: diagnostics.requestId,
-                    messageSummary: Flashcards.errorMessage(error: error)
-                )
-            )
-        )
-    }
-
-    @discardableResult
-    func captureCloudSyncFailureIfNeeded(
-        error: Error,
-        linkedSession: CloudLinkedSession,
-        fallbackCloudState: CloudAccountState?,
-        trigger: CloudSyncTrigger,
-        action: String
-    ) -> Bool {
-        if self.cloudCredentialRecoveryState?.reason == .linkedWorkspaceUnavailable,
-            let localStoreError = error as? LocalStoreError,
-            case .validation(let message) = localStoreError,
-            message == localizedCloudCredentialRecoveryBlockedMessage(reason: .linkedWorkspaceUnavailable) {
-            return false
-        }
-        if isLinkedWorkspaceUnavailableCloudSyncResponse(
-            error: error,
-            linkedSession: linkedSession,
-            cloudSettings: self.cloudSettings
-        ) {
-            return false
-        }
-        guard self.shouldCaptureCloudSyncFailure(error: error, trigger: trigger) else {
-            return false
-        }
-
-        self.captureCloudSyncFailure(
-            error: error,
-            linkedSession: linkedSession,
-            fallbackCloudState: fallbackCloudState,
-            action: action,
-            captureContext: trigger.technicalErrorCaptureContext
-        )
-        return true
-    }
-
-    private func shouldCaptureCloudSyncFailure(error: Error, trigger: CloudSyncTrigger) -> Bool {
-        if isRequestCancellationError(error: error) {
-            return false
-        }
-        if self.blockedCloudIdentityConflictMessage(error: error) != nil {
-            return false
-        }
-        if isRetryableNetworkTransportFailure(error: error) {
-            return false
-        }
-        if trigger.capturesTechnicalFailures {
-            return true
-        }
-
-        return self.isCloudAccountDeletedError(error)
     }
 
     func runLinkedSync(linkedSession: CloudLinkedSession) async throws -> CloudSyncResult {
@@ -1017,20 +749,6 @@ extension FlashcardsStore {
         return false
     }
 
-    private func enqueueBackgroundSyncVisibleChangeBannerIfNeeded(
-        bootstrapRefreshOutcome: BootstrapSnapshotRefreshOutcome
-    ) {
-        guard self.currentVisibleTab == .cards else {
-            return
-        }
-        guard bootstrapRefreshOutcome.workspaceChanged
-            || bootstrapRefreshOutcome.cardsChanged else {
-            return
-        }
-
-        self.enqueueTransientBanner(banner: makeCardsUpdatedFromCloudBanner())
-    }
-
     private func processMediaUploadTransfersAfterCloudSync(linkedSession: CloudLinkedSession) async {
         guard let database = self.database,
               let cloudSyncService = self.dependencies.cloudSyncService,
@@ -1053,79 +771,4 @@ extension FlashcardsStore {
         }
     }
 
-    private func captureMediaUploadTransferProcessingFailure(error: Error, linkedSession: CloudLinkedSession) {
-        if isSilentlyIgnorableNetworkTransportFailure(error: error) {
-            return
-        }
-
-        let diagnostics = cloudSyncFailureDiagnostics(error: error)
-        FlashcardsObservability.captureSilentFailure(
-            error: error,
-            scope: IOSObservationScope(
-                feature: .cloudSync,
-                userId: linkedSession.userId,
-                workspaceId: linkedSession.workspaceId,
-                requestId: diagnostics.requestId,
-                clientRequestId: nil,
-                sessionId: nil,
-                runId: nil,
-                cloudState: self.cloudSettings?.cloudState,
-                configurationMode: linkedSession.configurationMode
-            ),
-            action: "media_upload_transfer_process",
-            stage: "after_cloud_sync",
-            statusCode: diagnostics.statusCode,
-            backendCode: diagnostics.backendCode,
-            requestId: diagnostics.requestId
-        )
-    }
-}
-
-private struct CloudFailureDiagnostics {
-    let statusCode: Int?
-    let backendCode: String?
-    let requestId: String?
-}
-
-private func cloudSyncFailureDiagnostics(error: Error) -> CloudFailureDiagnostics {
-    if let syncError = error as? CloudSyncError {
-        switch syncError {
-        case .invalidResponse(let details, let statusCode):
-            return CloudFailureDiagnostics(
-                statusCode: statusCode,
-                backendCode: details.code,
-                requestId: details.requestId
-            )
-        case .invalidBaseUrl:
-            return CloudFailureDiagnostics(statusCode: nil, backendCode: nil, requestId: nil)
-        }
-    }
-
-    if let authError = error as? CloudAuthError {
-        switch authError {
-        case .invalidResponse(let details, let statusCode):
-            return CloudFailureDiagnostics(
-                statusCode: statusCode,
-                backendCode: details.code,
-                requestId: details.requestId
-            )
-        case .invalidBaseUrl, .invalidResponseBody:
-            return CloudFailureDiagnostics(statusCode: nil, backendCode: nil, requestId: nil)
-        }
-    }
-
-    if let guestAuthError = error as? GuestCloudAuthError {
-        switch guestAuthError {
-        case .invalidResponse(let details, let statusCode):
-            return CloudFailureDiagnostics(
-                statusCode: statusCode,
-                backendCode: details.code,
-                requestId: details.requestId
-            )
-        case .invalidBaseUrl, .invalidResponseBody:
-            return CloudFailureDiagnostics(statusCode: nil, backendCode: nil, requestId: nil)
-        }
-    }
-
-    return CloudFailureDiagnostics(statusCode: nil, backendCode: nil, requestId: nil)
 }
